+++ mobile_api/src/configs.rs
@@ -9,16 +9,12 @@
 //! reset.
 
 use crate::error::Result;
-use crate::security::{SecurityKey, SRNG};
-use crate::{device_config_path, device_info_path, sifis_home_path};
+use crate::security::SecurityKey;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::fs;
 use std::path::{Path, PathBuf};
 use uuid::Uuid;
-
-#[cfg(test)]
-mod tests;
 
 /// Smart Device Configuration
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq, JsonSchema, Serialize)]
@@ -43,14 +39,6 @@ impl DeviceConfig {
         &self.dht_shared_key
     }
 
-    /// Load from the default location
-    ///
-    /// This Convenience function tries to load a configuration file from
-    /// the location returned by the [device_config_path].
-    pub fn load() -> Result<DeviceConfig> {
-        Self::load_from(&device_config_path())
-    }
-
     /// Load from file
     ///
     /// Tries to load and parse configuration from the given *file* path.
@@ -62,14 +50,6 @@ impl DeviceConfig {
     /// Borrow device name
     pub fn name(&self) -> &str {
         &self.name
-    }
-
-    /// Write config to the default location.
-    ///
-    /// This convenience function tries to write configuration
-    /// to the file path given by the [device_config_path].
-    pub fn save(&self) -> Result<()> {
-        self.save_to(&device_config_path())
     }
 
     /// Save to file
@@ -113,23 +93,8 @@ pub struct DeviceInfo {
 }
 
 impl DeviceInfo {
-    /// Create a new device information
-    ///
-    /// Product name is required, other information is automatically generated.
-    pub fn new(product_name: String) -> Result<DeviceInfo> {
-        let srng = SRNG::new();
-        let mut private_key_file = sifis_home_path();
-        private_key_file.push("private.pem");
-        Ok(DeviceInfo {
-            authorization_key: srng.generate_key()?,
-            private_key_file,
-            product_name,
-            uuid: srng.generate_uuid()?,
-        })
-    }
-
     /// Create a new device information from known values
-    pub fn from(
+    pub fn new(
         product_name: String,
         authorization_key: SecurityKey,
         private_key_file: PathBuf,
@@ -148,28 +113,12 @@ impl DeviceInfo {
         &self.authorization_key
     }
 
-    /// Load from the default location
-    ///
-    /// This Convenience function tries to load a information file from
-    /// the location returned by the [device_info_path].
-    pub fn load() -> Result<DeviceInfo> {
-        Self::load_from(&device_info_path())
-    }
-
     /// Load from file
     ///
     /// Tries to load and parse device information from the given *file* path.
     pub fn load_from(file: &Path) -> Result<DeviceInfo> {
         let info_json = fs::read_to_string(file)?;
         Ok(serde_json::from_str::<DeviceInfo>(&info_json)?)
-    }
-
-    /// Write info to the default location.
-    ///
-    /// This convenience function tries to write information
-    /// to the file path given by the [device_info_path].
-    pub fn save(&self) -> Result<()> {
-        self.save_to(&device_info_path())
     }
 
     /// Save to file
@@ -224,5 +173,121 @@ impl DeviceInfo {
             true => serde_json::to_string_pretty(&self)?,
             false => serde_json::to_string(&self)?,
         })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::SifisHome;
+    use uuid::uuid;
+
+    const TEST_KEY_A: SecurityKey = SecurityKey::from_bytes([
+        0xf0, 0xe1, 0xd2, 0xc3, 0xb4, 0xa5, 0x96, 0x87, 0x78, 0x69, 0x5a, 0x4b, 0x3c, 0x2d, 0x1e,
+        0x0f, 0x0f, 0x1e, 0x2d, 0x3c, 0x4b, 0x5a, 0x69, 0x78, 0x87, 0x96, 0xa5, 0xb4, 0xc3, 0xd2,
+        0xe1, 0xf0,
+    ]);
+
+    const TEST_KEY_B: SecurityKey = SecurityKey::from_bytes([
+        0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e,
+        0x0f, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d,
+        0x1e, 0x1f,
+    ]);
+
+    pub const TEST_UUID: Uuid = Uuid::from_bytes([
+        0x12, 0x3e, 0x45, 0x67, 0xe8, 0x9b, 0x12, 0xd3, 0xa4, 0x56, 0x42, 0x66, 0x14, 0x17, 0x40,
+        0x00,
+    ]);
+
+    #[test]
+    fn test_device_config() {
+        // Testing constructor and getters
+        let mut config = DeviceConfig::new(TEST_KEY_A, "Test config".to_string());
+        assert_eq!(config.dht_shared_key(), &TEST_KEY_A);
+        assert_eq!(config.name(), "Test config");
+
+        // Testing setters and getters
+        config.set_dht_shared_key(TEST_KEY_B);
+        config.set_name(String::from("New name"));
+        assert_eq!(config.dht_shared_key(), &TEST_KEY_B);
+        assert_eq!(config.name(), "New name");
+    }
+
+    #[test]
+    fn test_device_config_serde() {
+        // Testing human readable with JSON
+        let config_a = DeviceConfig::new(SecurityKey::new().unwrap(), String::from("Test device"));
+        let json = serde_json::to_string(&config_a).unwrap();
+        let config_b = serde_json::from_str::<DeviceConfig>(&json).unwrap();
+        assert_eq!(config_a, config_b);
+
+        // Testing binary with MessagePack
+        let buf = rmp_serde::to_vec(&config_a).unwrap();
+        let config_b = rmp_serde::from_slice::<DeviceConfig>(&buf).unwrap();
+        assert_eq!(config_a, config_b);
+    }
+
+    #[test]
+    fn test_device_info() {
+        // Testing construction with SifisHome
+        let sifis_home = SifisHome::new();
+        let mut expected_private_key_file = PathBuf::from(sifis_home.home_path());
+        expected_private_key_file.push("private.pem");
+        let device = sifis_home.new_info("Test Device".to_string()).unwrap();
+        assert!(!device.authorization_key().is_null());
+        assert_eq!(device.private_key_file(), &expected_private_key_file);
+        assert_eq!(device.product_name(), "Test Device");
+        assert_eq!(device.uuid().get_version_num(), 7);
+
+        // Testing constructor and getters
+        let test_private_key = "/tmp/test/private.key";
+        let mut device = DeviceInfo::new(
+            "Test Device".to_string(),
+            TEST_KEY_A,
+            PathBuf::from(test_private_key),
+            TEST_UUID,
+        );
+        assert_eq!(device.authorization_key(), &TEST_KEY_A);
+        assert_eq!(device.private_key_file(), Path::new(test_private_key));
+        assert_eq!(device.product_name(), "Test Device");
+        assert_eq!(device.uuid(), &TEST_UUID);
+
+        // Testing setters and getters
+        let new_uuid = uuid!("5f8b3c30-ec2f-4228-af3b-dde564985e60");
+        let new_private_key = "/tmp/test/private.pem";
+        device.set_authorization_key(TEST_KEY_B);
+        device.set_private_key_file(PathBuf::from(&new_private_key));
+        device.set_product_name("New name".to_string());
+        device.set_uuid(new_uuid);
+        assert_eq!(device.authorization_key(), &TEST_KEY_B);
+        assert_eq!(device.private_key_file(), Path::new(new_private_key));
+        assert_eq!(device.product_name(), "New name");
+        assert_eq!(device.uuid(), &new_uuid);
+    }
+
+    #[test]
+    fn test_device_info_serde() {
+        let sifis_home = SifisHome::new();
+
+        // Testing human readable with JSON
+        let info_a = sifis_home.new_info(String::from("Test device")).unwrap();
+        let json = serde_json::to_string(&info_a).unwrap();
+        let info_b = serde_json::from_str::<DeviceInfo>(&json).unwrap();
+        assert_eq!(info_a, info_b);
+
+        // Testing binary with MessagePack
+        let buf = rmp_serde::to_vec(&info_a).unwrap();
+        let config_b = rmp_serde::from_slice::<DeviceInfo>(&buf).unwrap();
+        assert_eq!(info_a, config_b);
+
+        // Both compact and pretty JSON should result identical DeviceInfo object
+        let compact_json = info_a.to_json(false).unwrap();
+        let pretty_json = info_a.to_json(true).unwrap();
+        assert_ne!(compact_json, pretty_json);
+        let info_b = serde_json::from_str::<DeviceInfo>(&compact_json).unwrap();
+        let info_c = serde_json::from_str::<DeviceInfo>(&pretty_json).unwrap();
+        assert_eq!(info_a, info_b);
+        assert_eq!(info_b, info_c);
+        assert_eq!(info_b, info_c);
     }
 }
