+++ mobile_api_server/src/api_common.rs
@@ -1,3 +1,5 @@
+//! Common implementations for API endpoints
+
 use rocket::serde::json::Json;
 use rocket::serde::Deserialize;
 use rocket::Responder;
@@ -30,6 +32,10 @@ pub struct ErrorResponse {
 }
 
 impl ErrorResponse {
+    /// Constructing `400 Bad Request` Response
+    ///
+    /// The `description` option allows custom description,
+    /// but a default description is used by giving a `None` value.
     pub fn bad_request(description: Option<&str>) -> Json<ErrorResponse> {
         Json(ErrorResponse {
             error: ErrorResponseContent {
@@ -42,6 +48,10 @@ impl ErrorResponse {
         })
     }
 
+    /// Constructing `404 Not Found` Response
+    ///
+    /// The `description` option allows custom description,
+    /// but a default description is used by giving a `None` value.
     pub fn not_found(description: Option<&str>) -> Json<ErrorResponse> {
         Json(ErrorResponse {
             error: ErrorResponseContent {
@@ -54,6 +64,10 @@ impl ErrorResponse {
         })
     }
 
+    /// Constructing `500 Internal Server Error` Response
+    ///
+    /// This response should be used only for unexpected errors.
+    /// The `description` should contain a message of what went wrong.
     pub fn internal_server_error(description: String) -> Json<ErrorResponse> {
         Json(ErrorResponse {
             error: ErrorResponseContent {
@@ -64,11 +78,14 @@ impl ErrorResponse {
         })
     }
 
+    /// Constructing `503 Service Unavailable` Response
+    ///
+    /// The `description` should contain a message of why server is busy.
     pub fn service_unavailable(description: &str) -> Json<ErrorResponse> {
         Json(ErrorResponse {
             error: ErrorResponseContent {
                 code: 503,
-                reason: "Internal Server Error".to_string(),
+                reason: "Service Unavailable".to_string(),
                 description: description.to_string(),
             },
         })
@@ -86,6 +103,9 @@ pub struct OkResponse {
 }
 
 impl OkResponse {
+    /// Constructor for `200 OK` Response
+    ///
+    /// Message should tell what was done.
     pub fn message(message: &'static str) -> Json<OkResponse> {
         Json(OkResponse {
             code: 200,
@@ -94,14 +114,20 @@ impl OkResponse {
     }
 }
 
+/// A general set of server responses
+///
+/// Some endpoints have their collection of server responses, but this set is used in many.
 #[derive(Responder)]
 pub enum OkErrorBusyResponse {
+    /// 200 OK
     #[response(status = 200, content_type = "json")]
     Ok(Json<OkResponse>),
 
+    /// 500 Internal Server Server
     #[response(status = 500, content_type = "json")]
     Error(Json<ErrorResponse>),
 
+    /// 503 Service Unavailable
     #[response(status = 503, content_type = "json")]
     Busy(Json<ErrorResponse>),
 }
@@ -118,6 +144,25 @@ impl OpenApiResponderInner for OkErrorBusyResponse {
     }
 }
 
+/// Responses Generator
+///
+/// This responses generator allows an efficient way to implement [OpenApiResponderInner] for
+/// responses. In addition, the function automatically adds descriptions for some known status
+/// response codes.
+///
+/// # Example
+/// ```rust
+/// impl OpenApiResponderInner for FactoryResetResponse {
+///     fn responses(gen: &mut OpenApiGenerator) -> rocket_okapi::Result<Responses> {
+///         make_json_responses(vec![
+///             (200, gen.json_schema::<OkResponse>(), None),
+///             (400, gen.json_schema::<ErrorResponse>(), None),
+///             (500, gen.json_schema::<ErrorResponse>(), None),
+///             (503, gen.json_schema::<ErrorResponse>(), None),
+///         ])
+///     }
+/// }
+/// ```
 pub fn make_json_responses(
     status_schema_description: Vec<(u16, SchemaObject, Option<&str>)>,
 ) -> rocket_okapi::Result<Responses> {
