+++ mobile_api_server/src/main.rs
@@ -16,9 +16,7 @@
 //! See more Rocket related configuration options from: [rocket#configuration]
 
 use crate::state::DeviceState;
-use mobile_api::configs::DeviceInfo;
-use mobile_api::error::ErrorKind;
-use mobile_api::{device_info_path, sifis_home_path};
+use mobile_api::SifisHome;
 use rocket::fs::{relative, FileServer};
 use rocket::{Build, Rocket};
 use rocket_okapi::rapidoc::{make_rapidoc, GeneralConfig, HideShowConfig, RapiDocConfig};
@@ -38,39 +36,24 @@ async fn main() -> ExitCode {
     if dotenv::dotenv().is_ok() {
         println!("Loaded environment variables from .env file");
     }
+
+    // Using default SifisHome
+    let sifis_home = SifisHome::new();
     println!(
         "SIFIS-Home path: {}",
-        &sifis_home_path()
+        sifis_home
+            .home_path()
             .to_str()
             .expect("Could not get SIFIS-Home path")
     );
 
-    // Try to load device info and use it to create device state
-    let device_info = match DeviceInfo::load() {
-        Ok(device_info) => device_info,
-        Err(error) => {
-            // Special message for file not found error
-            if let ErrorKind::IoError(io_error) = error.kind() {
-                if io_error.kind() == std::io::ErrorKind::NotFound {
-                    eprintln!(
-                        "Device information file {:?} not found.",
-                        device_info_path()
-                    );
-                    eprintln!("You can use create_device_info application to create it.");
-                    return ExitCode::FAILURE;
-                }
-            };
-
-            // Error message for any other error
-            eprintln!(
-                "Could not load device information file: {:?}",
-                device_info_path()
-            );
-            eprintln!("{}", error);
+    let device_state = match DeviceState::new(sifis_home) {
+        Ok(device_state) => device_state,
+        Err(message) => {
+            eprintln!("{}", message);
             return ExitCode::FAILURE;
         }
     };
-    let device_state = DeviceState::new(device_info);
 
     let launch_result = build_rocket(device_state).launch().await;
